user-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
strawberry-graphql[fastapi]==0.219.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
python-multipart==0.0.6
cryptography==41.0.7
